Onboarding now fully works.
When a new user is detected, they will be asked to enter in their name and summer bridge group. From there, the GameObject that has all the blurbs and script to run the onboarding process is loaded and the the user can skip or go through it. Cleanup is ensured in both quiting as well as finishing the tutorial.