seperating components in search and scroll bar
Before search bar was built into the script handling the campaign screen. Since the quest screen could also use a search bar, the SearchBarComponent has been created to work with any screen with input field and container. Also added ResetOnScroll that resests the scrollview when the gameobject it's attached to gets reset.