fixed bug issues with screen
Redeem code screen was being destroyed when exiting it which caused error trying to open it up again. Issue when user exited the quest screen via the icons, causing the instructions screen to be stuck as inactive. Now ResetScreen() in QuestDetailScreen makes sure that when the icon button is pressed, it makes the intro screen visible again.