Renamed PlayerProgressMananger to StatManager
PlayerPorgressManager was intended to act as an inbetween app functions and the PlayerManager itself. It was removed during the process of translating the prototype into the figma build. Now the goal is add the inbetween again. This would also help as the PlayerManger is also acting as a subject for updating the player card when it shouldn't