fixing stats mananger and fixed issue with chest opening
Stats mananger is still broken but an issue with chest opening has been fixed. Before, it destory itself after two seconds. This was using the dstroy method. This had an issue with it destroying before the player had clicked to open the chest. Now ChestOpen uses a coroutine to wait for two seconds after the button has been pressed to ensure it doesn't destroy itself before the chest is opened.