Fixed tutorial / onboarding section
onboarding had issues after adjusting the campaign overview scren. Before, the tiles were out of order but now they are ordered by title. But the onboarding put the wrong title on the spotlight. Now it's been fixed. After changing how the Chest Open screen closes itself, it caused issues with the tutoral screen as its not using the RewardsScreenConstructor. Issue was fixed by activating the GetChestRewards directly on the button press of the chest found page.