fixed tutorial screens not filling the entire device dimensions
After the the onboarding finishes, the user gets to open a free chest to see it. The issue came on the android phone where the top and bottom of the screen did not get filled. This was due to the TutorialScreen gameobject containing the screens not being set to fill teh screen so it would leave it empty. Now this has been fixed.